Rename ArchComponents to AndroidLifecycle.
